+++ Questionable/Windows/UiUtils.cs
@@ -12,13 +12,14 @@ internal sealed class UiUtils(
     QuestData questData,
     ITextureProvider textureProvider)
 {
-    public (Vector4 Color, FontAwesomeIcon Icon, string Status) GetQuestStyle(ElementId elementId)
+    public unsafe (Vector4 Color, FontAwesomeIcon Icon, string Status) GetQuestStyle(ElementId elementId)
     {
         string lockedReason = string.Empty;
+        IQuestInfo? qInfo = null;
         questFunctions.prereqCache.TryGetValue(elementId.Value, out var prereqValue);
         if (questFunctions.IsQuestLocked(elementId) is (bool isLocked, string[] reasons) && isLocked)
             lockedReason = string.Join("\n  ", reasons);
-        else if (questData.TryGetQuestInfo(elementId, out var qInfo) && qInfo is IQuestInfo questInfo &&
+        else if (questData.TryGetQuestInfo(elementId, out qInfo) && qInfo is IQuestInfo questInfo &&
                 questInfo.PreviousQuestJoin is EQuestJoin.All)
         {
             List<string> preReasons = [];
@@ -52,6 +53,8 @@ internal sealed class UiUtils(
             return (QstTheme.Danger, FontAwesomeIcon.Times, $"{_L("Locked")}:\n  {lockedReason}");
         if (prereqValue == null)
             return (QstTheme.Info, FontAwesomeIcon.QuestionCircle, _L("Available(?)"));
+        if (qInfo != null && PlayerState.Instance()->CurrentLevel < qInfo.Level)
+            return (QstTheme.Danger, FontAwesomeIcon.Times, _L("Low level"));
 
         return (QstTheme.Amber, FontAwesomeIcon.Running, _L("Available"));
     }