+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Feature: Quest icons in journal progress (can be toggled in General > UI) -alydev
+  - Feature: quest icons in more places -alydev